requirements.txt
discord.py
python-dotenv
python-dateutil
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
aiohttp